Use around_filter instead of after_filter
Fixes https://github.com/mobalean/event_tracker/issues/7

Since event_tracker injects its HTML into the head tag with an
after_filter, a request that is halted in a before_filter and rendered
means that any other before_filter and after_filter methods won't be
called.

This introduces the problem where you cannot track metrics on any page
that is rendered in a before_filter (that is halted, of course).

This commit fixes the issue by wrapping the request in an around_filter
which is always called.